pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stagecoachml"
version = "0.1.0"
description = "A powerful machine learning pipeline orchestration framework"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Gaurav Sood", email = "contact@gsood.com"}
]
maintainers = [
    {name = "Gaurav Sood", email = "contact@gsood.com"}
]
keywords = [
    "machine learning",
    "ml pipeline",
    "data science",
    "orchestration",
    "workflow",
    "automation",
    "model training",
    "feature engineering",
    "dag",
    "etl"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed"
]
dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "networkx>=3.0",
    "joblib>=1.3.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "pytest-mock>=3.12.0",
    "pytest-timeout>=2.2.0",
    "pytest-benchmark>=4.0.0",
    "hypothesis>=6.90.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
    "types-pyyaml>=6.0.0",
    "pandas-stubs>=2.0.0"
]
docs = [
    "sphinx>=7.2.0",
    "furo>=2023.9.10",
    "sphinx-autodoc-typehints>=1.25.0",
    "myst-parser>=2.0.0",
    "sphinx-copybutton>=0.5.2",
    "sphinx-design>=0.5.0",
    "nbsphinx>=0.9.3"
]
viz = [
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
    "graphviz>=0.20.0"
]
all = [
    "stagecoachml[dev,docs,viz]"
]

[project.urls]
Homepage = "https://github.com/finite-sample/stagecoachml"
Documentation = "https://finite-sample.github.io/stagecoachml/"
Repository = "https://github.com/finite-sample/stagecoachml"
"Bug Tracker" = "https://github.com/finite-sample/stagecoachml/issues"
Changelog = "https://github.com/finite-sample/stagecoachml/blob/main/CHANGELOG.md"

[project.scripts]
stagecoach = "stagecoachml.cli:app"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/docs",
    "README.md",
    "LICENSE",
    "pyproject.toml"
]

[tool.hatch.build.targets.wheel]
packages = ["src/stagecoachml"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]
extend-include = ["*.ipynb"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
    "PTH",  # flake8-use-pathlib
    "RUF",  # ruff-specific rules
    "N",    # pep8-naming
    "D",    # pydocstyle
    "TID",  # flake8-tidy-imports
    "TCH",  # flake8-type-checking
    "PL",   # pylint
    "PT",   # flake8-pytest-style
]
ignore = [
    "D100", # Missing docstring in public module
    "D101", # Missing docstring in public class
    "D102", # Missing docstring in public method
    "D103", # Missing docstring in public function
    "D104", # Missing docstring in public package
    "D105", # Missing docstring in magic method
    "D107", # Missing docstring in __init__
    "D203", # 1 blank line required before class docstring
    "D213", # Multi-line docstring summary should start at the second line
    "PLR0913", # Too many arguments
    "PLR2004", # Magic value used in comparison
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "D104"]
"tests/*" = ["ARG", "D", "PLR2004", "S101"]
"docs/*" = ["D"]

[tool.ruff.lint.isort]
known-first-party = ["stagecoachml"]

[tool.ruff.format]
docstring-code-format = true

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
no_implicit_reexport = true
namespace_packages = true
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "sklearn.*",
    "joblib.*",
    "networkx.*",
    "yaml.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
addopts = [
    "-ra",
    "--strict-markers",
    "--cov=stagecoachml",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-branch",
    "--cov-fail-under=80",
    "-n=auto",
    "--tb=short"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "benchmark: marks tests as benchmarks"
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning"
]

[tool.coverage.run]
branch = true
source = ["src/stagecoachml"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/cli.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    "@overload"
]
precision = 2
show_missing = true